I get this error on line 112 of MainPage.xaml.cs: Argument 1: cannot convert from 'double' to 'float'
I get this warning on line 7 of NeighborInfo.cs: Non-nullable property 'Color' must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring the property as nullable.

+++ GridNomads/MainPage.xaml.cs
@@ -109,7 +109,7 @@ public partial class MainPage : ContentPage
         {
             var excitementLevel = nomad.ExcitementLevel;
             cells[(nomad.Row, nomad.Column)].BackgroundColor =
-                nomad.Color.WithLuminosity(0.5f + 0.3f * excitementLevel); // Adjust brightness based on excitement
+                nomad.Color.WithLuminosity((float)(0.5 + 0.3 * excitementLevel)); // Explicit cast to float
         }
     }
